requirements.txt
Django==1.9.6
Markdown==2.6.5
Pillow==3.2.0
PyYAML==3.11
argparse==1.2.1
bleach==1.4.3
csscompressor==0.9.4
django-admin-sortable2==0.6.3
django-appconf==1.0.2
django-autoslug==1.9.3
django-markitup==2.3.1
django-pipeline==1.6.8
-e git+https://github.com/mysociety/django-pipeline-csscompressor@6a88fde5045d68746f1c7ee816deda86615c5581#egg=django_pipeline_csscompressor-master
django-user-accounts==1.3.1
html5lib==0.9999999
libsass==0.11.1
ply==3.4
psycopg2==2.6.1
pytz==2016.4
six==1.10.0
slimit==0.8.1
sorl-thumbnail==12.4a1
unicodecsv==0.14.1
wsgiref==0.1.2
